Доработка для прохождения тестов Postman

+++ ewm-service/src/main/java/ru/practicum/event/dto/EventFullDto.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.practicum.category.dto.CategoryDto;
-import ru.practicum.location.Location;
+import ru.practicum.event.model.Location;
 import ru.practicum.user.dto.UserShortDto;
 
     @NoArgsConstructor
@@ -37,6 +37,6 @@ import ru.practicum.user.dto.UserShortDto;
         private String state;
         @NotNull
         private String title;
-        private Integer views;
+        private Long views;
     }
